+++ src/kompos/runners/compile.py
@@ -177,10 +177,7 @@ class CompileRunner(GenericRunner):
         super(CompileRunner, self).__init__(kompos_config, config_path, execute, RUNNER_TYPE)
 
     def run_configuration(self, args):
-        self.validate_runner = False
-        self.ordered_compositions = False
-        self.reverse = False
-        self.generate_output = False
+        self.configure_passive()
 
     def run(self, args, extra_args):
         self.run_configuration(args)